refactor: Move file nodes to individual files
Implements ML-597
